requirements.txt
Flask~=3.1.0
flask-restx~=1.3.0
Flask-JWT-Extended~=4.7.1
Flask-Cors~=5.0.0
PyJWT~=2.10.1
SQLAlchemy~=2.0.38
Werkzeug~=3.1.3
langchain~=0.3.19
langchain-openai~=0.3.6
langchain-community~=0.3.18
langchain-core~=0.3.37
pydantic~=2.10.6
python-dotenv~=1.0.1
pandas
numpy
matplotlib
tiktoken
faiss-cpu
pymysql
cryptography